Python scripts
Nelder_Mead_Method.py = a script where you are free to choose the location of the celestial bodies, their size, and mass, and where the function Nelder-Mead takes as input only the 3 starting vertices(the space is 2D, so the simplex is a triangle). Attention: the algorithm is not bounded!

Nelder_Mead_For_Sector.py = same script as above except that now the map has been sectioned into small rectangular sectors(grid) and the Nelder-Mead algorithm has been restricted to find the minimum only for the sector, and store the coordinates, as well as the value, in a 2D array. The algorithm is bounded. Depending on the sector, the 3 starting vertices are selected to be in the middle of the rectangular sector for which we are calculating. There is a penalty introduced in the force calculation function, where if the new vertex is outside of the bounds, it will return a very large value and force the algorithm to not step outside of the sector. You can select in how many square sectors you would like to divide the map, you can also select how many optimization steps you would like the algorithm to make.

Dijkstra_alg.py = A failed attempt at trying to combine the code above with a Weighted Dijkstra Algorithm so as to find the optimal path from point A(x_A,y_A) to point B(x_B,y_B) based on coefficients for the force at the different waypoints and the distance to B.

Random_jumping.py = A random jumping algorithm that tries to find the global minimum for N number of random points